+++ libs/frontend/modules/graphql/src/services/ranking-system.service.ts
@@ -74,7 +74,6 @@ export class RankingSystemService {
     switchMap((saved) => this._loadSystem(saved)),
   );
 
-  //sources
   sources$ = merge(
     this.servicesLoaded$.pipe(
       map((rankingSystem) => ({